Worked on APIs ensuring smoooth communication between the frontend and the backend

+++ app/routes/issues.py
@@ -3,6 +3,7 @@ from app.models.issue import Issue
 from app.models import db
 from flask_jwt_extended import get_jwt_identity, jwt_required
 
+
 issues_bp = Blueprint('issues', __name__)
 
 
@@ -18,7 +19,7 @@ def create_issue():
 
     # Ensure data is present
     if not data:
-        return jsonify({'error:' 'Missing JSON data'}), 400
+        return jsonify({'error': 'Missing JSON data'}), 400
     
     required_fields = ['title', 'location','description']
     missing_fields = [field for field in required_fields if field not in data]
@@ -45,6 +46,7 @@ def create_issue():
 
 # Get all issues
 @issues_bp.route('/', methods=['GET'])
+@issues_bp.route('', methods=['GET'])
 def get_issues():
     issues = Issue.query.all()
     issues_list = [{
@@ -73,21 +75,30 @@ def get_issue(issue_id):
     })
 
 # update a new issue
-@issues_bp.route('/<int:issue_id>', methods=['PUT'])
+@issues_bp.route('/<int:issue_id>', methods=['PUT', 'DELETE'])
+@jwt_required()
 def update_issue(issue_id):
-    data = request.get_json()
+    current_user_id = get_jwt_identity()
     issue = Issue.query.get_or_404(issue_id)
-    issue.title = data.get('title', issue.title)
-    issue.location = data.get('location', issue.location)
-    issue.description = data.get('description', issue.description)
-    db.session.commit()
-    return jsonify({'message': 'Issue updated successfully'})
 
-
-# Delete an issue
-@issues_bp.route('/<int:issue_id>', methods=['DELETE'])
-def delete_issue(issue_id):
-    issue = Issue.query.get_or_404(issue_id)
-    db.session.delete(issue)
-    db.session.commit()
-    return jsonify({'message': 'Issue deleted successfully'})
+    # Ensure user is the owner of the issue
+    if issue.user_id != current_user_id:
+        return jsonify({'error': 'Unauthorized to modify the issue'}), 403
+    
+    if request.method == 'PUT':
+        data = request.get_json()
+        if not data:
+            return jsonify({'error': 'Missing JSON data'}), 400
+        
+        issue.title = data.get('title', issue.title)
+        issue.location = data.get('location', issue.location)
+        issue.description = data.get('description', issue.description)
+        db.session.commit()
+        return jsonify({'message': 'Issue updated successfully'})
+ 
+    
+    elif request.method == 'DELETE':
+        db.session.delete(issue)
+        db.session.commit()
+        return jsonify({'message': 'Issue deleted successfully'})
+
